+++ tools/python/check_paradox_html/miraitranslator_com_register_forgot_php.txt
@@ -1,0 +1,73 @@
+<html ng-app="Register">
+<head>
+<meta charset="utf-8">
+<title>Mirai Translator</title>
+<link href="/contents/mt/images/favicon.ico" rel="icon" type="image/vnd.microsoft.icon">
+<link href="css/reset.css" rel="stylesheet">
+<link href="css/layout.css" rel="stylesheet">
+<link href="css/pulldown.css" rel="stylesheet">
+<link href="../iconfont/style.css" rel="stylesheet">
+<link href="../theme/theme-white.css" rel="stylesheet">
+<script src="js/prefixfree.min.js"></script>
+<script src="js/angular.js" type="text/javascript"></script>
+</link></link></link></link></link></link></meta></head>
+<body>
+<header class="loginHeader">
+<!-- <div id="pulldown">
+                  <span class="icon-language i-language"></span>
+                  <div class="custom-select-wrapper">
+                        <div class="custom-select sources">
+                              <span class="custom-select-trigger">Language</span>
+                              <div class="custom-options">
+                                    <span class="custom-option " data-value="English">English</span>
+                                    <span class="custom-option " data-value="Japanese">Japanese</span>
+                                    <span class="custom-option " data-value="Chinese">Chinese</span>
+                              </div>
+                        </div>
+                  </div>
+            </div> -->
+</header>
+<main>
+<p id="logo"><img src="/contents/mt/images/MTlogo.png"/></p>
+<h1>Forgot your password?</h1>
+<p class="margin-bottom40">We'll send you an email with a reset link.</p>
+<form method="post" name="forgot" novalidate="">
+<dl>
+<dt>Email:</dt>
+<dd>
+<input name="email" ng-model="emailModel" ng-required="true" required="" type="email">
+<span ng-if="forgot.email.$touched" ng-show="forgot.email.$error.required"><br>
+<span class="text-error">Require Email</span>
+</br></span>
+<span ng-if="forgot.email.$touched" ng-show="forgot.email.$error.email"><br>
+<span class="text-error">Input email format</span>
+</br></span>
+</input></dd>
+</dl>
+<div class="submit">
+<button ng-disabled="forgot.email.$invalid" onclick="document.forgot.submit(); return false;">Reset password</button>
+</div>
+</form>
+<script language="javascript">
+
+      angular.module("Register", [])
+        .directive("match", ["$parse", function($parse) {
+           return {
+            require: 'ngModel',
+            link: function(scope, elem, attrs, ctrl) {
+              scope.$watch(function() {
+                var target = $parse(attrs.match)(scope);
+                return !ctrl.$modelValue || target === ctrl.$modelValue;
+              }, function(currentValue) {
+                ctrl.$setValidity('mismatch', currentValue);
+              });
+            }
+          }
+        }]);
+    </script>
+</main>
+<!-- <footer class="base_footer">© Mirai Translate 2016</footer> -->
+<script src="https://ajax.googleapis.com/ajax/libs/jquery/2.1.4/jquery.min.js"></script>
+<script src="js/index.js"></script>
+</body>
+</html>
