+++ tools/python/check_paradox_html/miraitranslator_com_register_signin_php.txt
@@ -1,0 +1,75 @@
+<html ng-app="Register">
+<head>
+<meta charset="utf-8">
+<title>Mirai Translator</title>
+<link href="/contents/mt/images/favicon.ico" rel="icon" type="image/vnd.microsoft.icon">
+<link href="css/reset.css" rel="stylesheet">
+<link href="css/layout.css" rel="stylesheet">
+<link href="css/pulldown.css" rel="stylesheet">
+<link href="../iconfont/style.css" rel="stylesheet">
+<link href="../theme/theme-white.css" rel="stylesheet">
+<script src="js/prefixfree.min.js"></script>
+<script src="js/angular.js" type="text/javascript"></script>
+</link></link></link></link></link></link></meta></head>
+<body>
+<header class="loginHeader">
+<!-- <div id="pulldown">
+                  <span class="icon-language i-language"></span>
+                  <div class="custom-select-wrapper">
+                        <div class="custom-select sources">
+                              <span class="custom-select-trigger">Language</span>
+                              <div class="custom-options">
+                                    <span class="custom-option " data-value="English">English</span>
+                                    <span class="custom-option " data-value="Japanese">Japanese</span>
+                                    <span class="custom-option " data-value="Chinese">Chinese</span>
+                              </div>
+                        </div>
+                  </div>
+            </div> -->
+</header>
+<main>
+<p id="logo">
+<img src="/contents/mt/images/MTlogo.png">
+</img></p>
+<h1>Mirai Translator</h1>
+<form method="post" name="signin" novalidate="">
+<dl>
+<dt>Email:</dt>
+<dd><input autocomplete="off" id="email" name="email" placeholder="Enter Email Address" type="email" value=""/></dd>
+<dt>Password:</dt>
+<dd><input id="password" name="password" placeholder="Enter Password" type="password" 　=""><div class="box300w"><a href="forgot.php">Forgot your password?</a></div></input></dd>
+<dt></dt>
+<dd>
+<input id="remember" name="remember" type="checkbox" value="remember"><label class="remember" for="remember">Remember me</label>
+<br>
+<p class="btn-default btn-submit margin-top16"><a href="javascript:void(0);" onclick="document.signin.submit(); return false;">Sign in</a></p>
+<br>
+</br></br></input></dd>
+</dl>
+</form>
+<ul class="description">
+<li>
+<a href="/contents/mt/pdf/Terms_of_Use.pdf" target="_blank">Term of Use</a>
+</li>
+<li>・</li>
+<li>
+<a href="/contents/mt/pdf/Privacy_Policy.pdf" target="_blank">Privacy Policy</a>
+</li>
+</ul>
+<!-- Enterでログイン -->
+<script>
+document.onkeydown = function (e){
+	var key_code = e.keyCode;
+	if(key_code == 13){
+		if($("#email").val() && $("#password").val()){
+			document.signin.submit();
+		}
+	}
+};
+</script>
+</main>
+<!-- <footer class="base_footer">© Mirai Translate 2016</footer> -->
+<script src="https://ajax.googleapis.com/ajax/libs/jquery/2.1.4/jquery.min.js"></script>
+<script src="js/index.js"></script>
+</body>
+</html>
